requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==5.2
django-extensions==4.1
django-import-export==4.3.7
narwhals==1.35.0
numpy==2.2.4
packaging==24.2
pandas==2.2.3
plotly==6.0.1
python-dateutil==2.9.0.post0
pytz==2025.2
six==1.17.0
sqlparse==0.5.3
tablib==3.8.0
tzdata==2025.2
